feat(DOS-FIX): onboarding empty state, sample dossier loader, archive UI
Make the first-run experience at /dossiers legible without a tutorial.
New users see a workflow explainer and two CTAs: 'Start a dossier'
(blank) and 'Load sample dossier' (seeds the curated Northgate Pharma
demo into their account).

Extract the per-dossier seed content from prisma/seed.ts into a
reusable buildSampleDossier(client, ownerId) helper so the same
content powers both the dev seed CLI and the user-facing loader. Slug
collisions auto-suffix so loading the sample twice creates two
separate dossiers rather than erroring.

Add a per-row dropdown menu to the dossier list with an Archive
action. Confirmation dialog matches NewDossierModal's native dialog
pattern. Wires existing archiveDossier server action; the dossier
query already filters by status='active' so archived rows disappear
on revalidate. No hard-delete; data is recoverable at the DB level.
