+++ src/server/lib/sample-dossier.ts
@@ -1,0 +1,679 @@
+import type { PrismaClient } from "@prisma/client";
+
+// Resolve highlight offsets from the real raw_text so edits to source copy
+// cannot silently desync the evidence-gutter bands in the reader.
+function findOffsets(
+  text: string,
+  quote: string,
+): { start_offset: number; end_offset: number } {
+  const start = text.indexOf(quote);
+  if (start < 0) {
+    throw new Error(
+      `Sample dossier integrity error: highlight quote not found in source text.\nQuote: ${JSON.stringify(quote.slice(0, 80))}...`,
+    );
+  }
+  return { start_offset: start, end_offset: start + quote.length };
+}
+
+/**
+ * Build the curated "Northgate Pharma" sample dossier under the given owner.
+ *
+ * Used by both the dev seed CLI (`prisma/seed.ts`) and the user-facing
+ * "Load sample dossier" server action so the demo content has one source of
+ * truth.
+ *
+ * Returns the new dossier's id so callers can redirect or report on it.
+ */
+export async function buildSampleDossier(
+  client: PrismaClient,
+  ownerId: string,
+): Promise<{ dossierId: string }> {
+  // ─── Tags ────────────────────────────────────────────────────────────────────
+  // Tags are unique by name and shared across users; upsert so repeated
+  // invocations don't collide.
+
+  const tagRegulatory = await client.tag.upsert({
+    where: { name: "regulatory" },
+    update: {},
+    create: { name: "regulatory" },
+  });
+
+  const tagFinancial = await client.tag.upsert({
+    where: { name: "financial" },
+    update: {},
+    create: { name: "financial" },
+  });
+
+  const tagManufacturing = await client.tag.upsert({
+    where: { name: "manufacturing" },
+    update: {},
+    create: { name: "manufacturing" },
+  });
+
+  // ─── Dossier ─────────────────────────────────────────────────────────────────
+  // Slug is uniqued per owner; suffix on collision so a user who loads the
+  // sample twice gets two separate dossiers rather than an error.
+
+  const baseSlug = "northgate-pharma-fda-warning";
+  let slug = baseSlug;
+  let n = 1;
+  while (await client.dossier.findFirst({ where: { owner_id: ownerId, slug } })) {
+    slug = `${baseSlug}-${n++}`;
+  }
+
+  const dossier = await client.dossier.create({
+    data: {
+      owner_id: ownerId,
+      title: "Northgate Pharma — FDA Warning Letter Investigation",
+      slug,
+      summary:
+        "Investigation into systemic manufacturing violations at Northgate Pharma's Raleigh facility and the company's regulatory exposure following a 2024 FDA warning letter.",
+      status: "active",
+      research_goal:
+        "Determine whether manufacturing violations are systemic, establish a timeline of internal knowledge, and assess total litigation and remediation exposure.",
+      priority: 1,
+    },
+  });
+
+  // ─── Sources ─────────────────────────────────────────────────────────────────
+
+  const fdaLetterText = `Dear Mr. Harrington,
+
+During our inspection of Northgate Pharmaceutical Inc., 4200 Research Triangle Pkwy, Raleigh, NC, conducted October 16–November 2, 2023, our investigator(s) identified significant violations of Current Good Manufacturing Practice (CGMP) regulations for finished pharmaceuticals, Title 21, Code of Federal Regulations, Parts 210 and 211.
+
+Specifically, we found:
+
+1. Your firm failed to establish adequate written procedures for production and process controls (21 CFR 211.100(a)). Batch records for Lot Nos. NGP-2023-0441 through NGP-2023-0489 lacked required in-process testing documentation.
+
+2. Your firm failed to maintain equipment in a clean and orderly manner and in good repair (21 CFR 211.67(a)). Sterile filling Line 3 exhibited visible particulate contamination residue from prior production runs. No documented cleaning verification was performed between October 4–12, 2023.
+
+3. Your firm failed to exercise appropriate controls over computer or related systems to assure that only authorized personnel institute changes in master production records (21 CFR 211.68(b)). Audit log analysis revealed 47 instances of data backdating between March 2022 and September 2023.
+
+We acknowledge receipt of your response dated November 30, 2023. After reviewing your response, we conclude it does not adequately address our concerns. Failure to promptly correct these violations may result in legal action without further notice, including, without limitation, seizure and injunction.`;
+
+  const sourceFdaLetter = await client.source.create({
+    data: {
+      dossier_id: dossier.id,
+      type: "pdf",
+      title: "FDA Warning Letter — Northgate Pharma (WL-2024-0108)",
+      author: "U.S. Food and Drug Administration",
+      publisher: "FDA",
+      published_at: new Date("2024-01-08"),
+      raw_text: fdaLetterText,
+      summary:
+        "FDA identifies three CGMP violations: inadequate batch documentation, equipment contamination, and 47 instances of audit log backdating over 18 months.",
+      credibility_rating: 5,
+      source_status: "reviewed",
+      tags: {
+        create: [
+          { tag_id: tagRegulatory.id },
+          { tag_id: tagManufacturing.id },
+        ],
+      },
+    },
+  });
+
+  const reutersText = `Shares of Northgate Pharmaceutical Inc. fell as much as 18% on Thursday after the company disclosed it had received an FDA warning letter citing data integrity failures and manufacturing control deficiencies at its Raleigh, North Carolina facility.
+
+The warning letter, dated January 8, 2024, is the company's second regulatory action in four years. In 2020, Northgate received a Form 483 observation letter citing similar documentation gaps at the same facility, raising questions about whether management took adequate corrective action.
+
+"This is not a first offense situation," said Elena Vasquez, a pharmaceutical regulatory consultant at Meridian Advisory Group. "The 2020 observations were a clear warning. If Northgate's board was paying attention, they knew the risk."
+
+Northgate's Chief Scientific Officer, Dr. Sandra Wei, issued a statement saying the company "respectfully disagrees with several characterizations in the letter" and has initiated a comprehensive remediation program with an outside consultant.
+
+Analysts at Morgan Stanley estimated the financial impact at between $40M and $120M depending on whether the FDA proceeds to a consent decree. The company's Raleigh facility accounts for approximately 34% of total revenue.`;
+
+  const sourceReuters = await client.source.create({
+    data: {
+      dossier_id: dossier.id,
+      type: "web_link",
+      title:
+        "Northgate Pharma shares drop 18% after FDA warning letter disclosed",
+      url: "https://example.com/reuters-northgate-fda-2024",
+      author: "Priya Chandrasekaran",
+      publisher: "Reuters",
+      published_at: new Date("2024-01-11"),
+      raw_text: reutersText,
+      summary:
+        "Reuters coverage of the FDA warning letter disclosure. Notes prior 2020 Form 483 at same facility, analyst estimates of $40M-$120M exposure, and CSO Dr. Sandra Wei's public response.",
+      credibility_rating: 4,
+      source_status: "reviewed",
+      tags: {
+        create: [{ tag_id: tagFinancial.id }, { tag_id: tagRegulatory.id }],
+      },
+    },
+  });
+
+  const memoText = `TO: Sandra Wei, CSO; David Harrington, CEO
+FROM: Raymond Chu, VP Quality Assurance
+DATE: June 14, 2022
+RE: Escalation — Line 3 Sterile Fill Data Integrity Concerns
+
+I am escalating a matter that Quality has flagged internally since Q1 2022 and that has not received adequate resolution.
+
+Our internal audit in May 2022 identified a pattern of retroactive entries in the LIMS system for Sterile Fill Line 3. Specifically, 12 batch records between January and April 2022 show timestamp anomalies consistent with post-hoc data entry rather than contemporaneous recording. This is a potential 21 CFR 211.68(b) violation.
+
+I raised this with the site Quality Director on March 18, 2022. No corrective action plan has been issued. Given the pattern and the timeline, I believe the company faces material regulatory risk if this is not addressed before the next FDA inspection cycle.
+
+I am formally requesting executive sign-off on a CAPA within 30 days.
+
+Raymond Chu`;
+
+  const sourceMemo = await client.source.create({
+    data: {
+      dossier_id: dossier.id,
+      type: "internal_memo",
+      title: "Internal QA Escalation Memo — Raleigh Facility (June 2022)",
+      author: "Raymond Chu, VP Quality Assurance",
+      published_at: new Date("2022-06-14"),
+      raw_text: memoText,
+      summary:
+        "Internal QA memo from June 2022 documenting data integrity concerns on Line 3 escalated to CSO and CEO. Chu notes the issue was first raised in March 2022 with no corrective action taken.",
+      credibility_rating: 5,
+      source_status: "reviewed",
+      tags: {
+        create: [
+          { tag_id: tagManufacturing.id },
+          { tag_id: tagRegulatory.id },
+        ],
+      },
+    },
+  });
+
+  const interviewText = `Q: How serious are audit log backdating findings in FDA warning letters?
+
+A: They're among the most serious findings we can make. Audit trail manipulation goes to the fundamental integrity of the GMP data system. It's not a paperwork error — it's a data integrity failure. The agency has been exceptionally aggressive on this since 2016.
+
+Q: Forty-seven instances over 18 months — does that suggest a systemic problem or isolated incidents?
+
+A: Forty-seven instances is not rogue behavior by a single employee. That pattern, sustained over 18 months, tells me there was either a cultural tolerance for this or it was tacitly encouraged. Either way, that's a management failure, not a QA failure.
+
+Q: What's the risk of a consent decree here?
+
+A: If the company had a prior 483 at the same site and documentation gaps weren't fully remediated, the probability of a consent decree is high. I'd put it above 50%. The financial range analysts are quoting — $40M to $120M — honestly, consent decrees for facilities this size can run higher if they include facility remediation requirements. I would not anchor to that lower number.`;
+
+  const sourceExpertInterview = await client.source.create({
+    data: {
+      dossier_id: dossier.id,
+      type: "pasted_text",
+      title: "Expert Interview — Dr. Marcus Obi, Former FDA Investigator",
+      author: "Dr. Marcus Obi",
+      published_at: new Date("2024-02-15"),
+      raw_text: interviewText,
+      summary:
+        "Interview with former FDA investigator Dr. Marcus Obi. Characterizes the backdating pattern as a management failure, not isolated misconduct. Consent decree probability assessed at above 50%. Skeptical of analyst $40M floor.",
+      credibility_rating: 4,
+      source_status: "reviewing",
+    },
+  });
+
+  // An unreviewed source demonstrates the capture → triage pipeline: a
+  // freshly-clipped piece of evidence that still needs the researcher's
+  // attention.
+  await client.source.create({
+    data: {
+      dossier_id: dossier.id,
+      type: "web_link",
+      title:
+        "Northgate 2023 Annual Shareholder Letter (excerpt): 'industry-leading quality culture'",
+      url: "https://example.com/northgate-2023-annual-letter",
+      author: "David Harrington, CEO",
+      publisher: "Northgate Pharmaceutical Inc.",
+      published_at: new Date("2023-03-22"),
+      raw_text: `To our shareholders,
+
+The past fiscal year reflected disciplined execution across every function of the business. Our Raleigh facility continues to operate under an industry-leading quality culture and remains central to our long-term growth thesis. Investments in quality systems and data integrity programs, made consistently over the past three years, distinguish Northgate from peer manufacturers facing regulatory headwinds.
+
+We are confident in the compliance posture of our manufacturing network as we enter 2024.
+
+— David Harrington, CEO`,
+      summary:
+        "CEO letter from March 2023 asserting an 'industry-leading quality culture' at the Raleigh facility — sits in direct tension with the contemporaneous internal QA escalation.",
+      credibility_rating: 3,
+      source_status: "unreviewed",
+      tags: { create: [{ tag_id: tagFinancial.id }] },
+    },
+  });
+
+  // ─── Highlights ───────────────────────────────────────────────────────────────
+
+  const hl1 = await client.highlight.create({
+    data: {
+      source_id: sourceFdaLetter.id,
+      quote_text:
+        "Audit log analysis revealed 47 instances of data backdating between March 2022 and September 2023.",
+      ...findOffsets(
+        fdaLetterText,
+        "Audit log analysis revealed 47 instances of data backdating between March 2022 and September 2023.",
+      ),
+      label: "stat",
+      annotation: "Core finding — 18-month window of systematic backdating.",
+    },
+  });
+
+  const hl2 = await client.highlight.create({
+    data: {
+      source_id: sourceFdaLetter.id,
+      quote_text:
+        "Sterile filling Line 3 exhibited visible particulate contamination residue from prior production runs. No documented cleaning verification was performed between October 4–12, 2023.",
+      ...findOffsets(
+        fdaLetterText,
+        "Sterile filling Line 3 exhibited visible particulate contamination residue from prior production runs. No documented cleaning verification was performed between October 4–12, 2023.",
+      ),
+      label: "evidence",
+      annotation: "Physical contamination distinct from data integrity issue.",
+    },
+  });
+
+  const hl3 = await client.highlight.create({
+    data: {
+      source_id: sourceReuters.id,
+      quote_text:
+        "In 2020, Northgate received a Form 483 observation letter citing similar documentation gaps at the same facility, raising questions about whether management took adequate corrective action.",
+      ...findOffsets(
+        reutersText,
+        "In 2020, Northgate received a Form 483 observation letter citing similar documentation gaps at the same facility, raising questions about whether management took adequate corrective action.",
+      ),
+      label: "evidence",
+      annotation: "Prior notice — company was on record as knowing the risk.",
+    },
+  });
+
+  const hl4 = await client.highlight.create({
+    data: {
+      source_id: sourceMemo.id,
+      quote_text:
+        "I raised this with the site Quality Director on March 18, 2022. No corrective action plan has been issued.",
+      ...findOffsets(
+        memoText,
+        "I raised this with the site Quality Director on March 18, 2022. No corrective action plan has been issued.",
+      ),
+      label: "evidence",
+      annotation:
+        "Internal escalation with no follow-through. March 2022 is 21 months before FDA inspection.",
+    },
+  });
+
+  const hl5 = await client.highlight.create({
+    data: {
+      source_id: sourceMemo.id,
+      quote_text:
+        "I am formally requesting executive sign-off on a CAPA within 30 days.",
+      ...findOffsets(
+        memoText,
+        "I am formally requesting executive sign-off on a CAPA within 30 days.",
+      ),
+      label: "quote",
+      annotation:
+        "Chu put leadership on notice in writing. Key document for knowledge timeline.",
+    },
+  });
+
+  const hl6 = await client.highlight.create({
+    data: {
+      source_id: sourceExpertInterview.id,
+      quote_text:
+        "Forty-seven instances is not rogue behavior by a single employee. That pattern, sustained over 18 months, tells me there was either a cultural tolerance for this or it was tacitly encouraged.",
+      ...findOffsets(
+        interviewText,
+        "Forty-seven instances is not rogue behavior by a single employee. That pattern, sustained over 18 months, tells me there was either a cultural tolerance for this or it was tacitly encouraged.",
+      ),
+      label: "quote",
+      annotation:
+        "Expert opinion framing this as a management/culture failure, not individual misconduct.",
+    },
+  });
+
+  const hl7 = await client.highlight.create({
+    data: {
+      source_id: sourceExpertInterview.id,
+      quote_text: "I would not anchor to that lower number.",
+      ...findOffsets(interviewText, "I would not anchor to that lower number."),
+      label: "counterpoint",
+      annotation:
+        "Expert explicitly pushes back on the $40M floor estimate used by analysts.",
+    },
+  });
+
+  // ─── Claims ───────────────────────────────────────────────────────────────────
+
+  const claimSystemic = await client.claim.create({
+    data: {
+      dossier_id: dossier.id,
+      statement:
+        "The manufacturing violations at Northgate's Raleigh facility were systemic, not isolated incidents.",
+      confidence: 85,
+      status: "supported",
+      notes:
+        "Supported by: 47 backdating instances over 18 months (FDA), expert characterization of a management/culture failure (Obi interview), and prior 2020 483 at same facility.",
+      highlights: {
+        create: [
+          { highlight_id: hl1.id },
+          { highlight_id: hl3.id },
+          { highlight_id: hl6.id },
+        ],
+      },
+      tags: { create: [{ tag_id: tagManufacturing.id }] },
+    },
+  });
+
+  const claimKnowledge = await client.claim.create({
+    data: {
+      dossier_id: dossier.id,
+      statement:
+        "Northgate's executive leadership had documented internal knowledge of data integrity risks at least 21 months before the FDA inspection.",
+      confidence: 90,
+      status: "supported",
+      notes:
+        "The June 2022 internal QA memo addressed to the CSO and CEO establishes executive-level knowledge. The March 2022 escalation to the Quality Director is an additional prior notice.",
+      highlights: {
+        create: [{ highlight_id: hl4.id }, { highlight_id: hl5.id }],
+      },
+      tags: { create: [{ tag_id: tagRegulatory.id }] },
+    },
+  });
+
+  await client.claim.create({
+    data: {
+      dossier_id: dossier.id,
+      statement:
+        "Total financial exposure from this FDA action is unlikely to be contained within the $40M analyst floor estimate.",
+      confidence: 65,
+      status: "open",
+      notes:
+        "Expert opinion and prior consent decree data suggest the lower bound is too optimistic. Need to pull comparable consent decree financials for context.",
+      highlights: {
+        create: [{ highlight_id: hl7.id }],
+      },
+      tags: { create: [{ tag_id: tagFinancial.id }] },
+    },
+  });
+
+  // ─── Entities ─────────────────────────────────────────────────────────────────
+
+  const entityNorthgate = await client.entity.create({
+    data: {
+      dossier_id: dossier.id,
+      name: "Northgate Pharmaceutical Inc.",
+      type: "company",
+      description:
+        "Mid-size specialty pharmaceutical manufacturer headquartered in Raleigh, NC. Subject of the FDA warning letter and primary subject of this investigation.",
+      aliases: ["Northgate Pharma", "NGP", "Northgate"],
+      importance: 10,
+      tags: {
+        create: [
+          { tag_id: tagRegulatory.id },
+          { tag_id: tagManufacturing.id },
+        ],
+      },
+    },
+  });
+
+  const entityWei = await client.entity.create({
+    data: {
+      dossier_id: dossier.id,
+      name: "Dr. Sandra Wei",
+      type: "person",
+      description:
+        "Chief Scientific Officer of Northgate Pharma. Named recipient of the June 2022 internal QA memo. Issued public statement disputing some FDA characterizations.",
+      aliases: ["Wei", "S. Wei"],
+      importance: 7,
+    },
+  });
+
+  const entityChu = await client.entity.create({
+    data: {
+      dossier_id: dossier.id,
+      name: "Raymond Chu",
+      type: "person",
+      description:
+        "VP of Quality Assurance at Northgate Pharma. Author of the June 2022 escalation memo. Key witness to internal knowledge timeline.",
+      aliases: ["Chu", "R. Chu"],
+      importance: 8,
+    },
+  });
+
+  const entityHarrington = await client.entity.create({
+    data: {
+      dossier_id: dossier.id,
+      name: "David Harrington",
+      type: "person",
+      description:
+        "Chief Executive Officer of Northgate Pharma. Named recipient of the June 2022 QA escalation memo and author of the March 2023 shareholder letter asserting an 'industry-leading quality culture.'",
+      aliases: ["Harrington", "D. Harrington"],
+      importance: 9,
+    },
+  });
+
+  // ─── Mentions ─────────────────────────────────────────────────────────────────
+  // Look up the unreviewed shareholder letter source by URL since we didn't
+  // hold a reference to it (it has no highlights or claims of its own yet).
+
+  const sourceShareholderLetter = await client.source.findFirstOrThrow({
+    where: {
+      dossier_id: dossier.id,
+      url: "https://example.com/northgate-2023-annual-letter",
+    },
+  });
+
+  await client.mention.create({
+    data: {
+      entity_id: entityNorthgate.id,
+      source_id: sourceFdaLetter.id,
+      highlight_id: hl1.id,
+      context_snippet:
+        "Northgate Pharmaceutical Inc. — primary subject of FDA Warning Letter WL-2024-0108.",
+    },
+  });
+
+  await client.mention.create({
+    data: {
+      entity_id: entityWei.id,
+      source_id: sourceMemo.id,
+      highlight_id: hl5.id,
+      context_snippet:
+        "Dr. Sandra Wei named as memo recipient alongside CEO David Harrington.",
+    },
+  });
+
+  await client.mention.create({
+    data: {
+      entity_id: entityChu.id,
+      source_id: sourceMemo.id,
+      highlight_id: hl4.id,
+      context_snippet:
+        "Raymond Chu authored the June 2022 escalation memo and signed the CAPA request.",
+    },
+  });
+
+  await client.mention.create({
+    data: {
+      entity_id: entityNorthgate.id,
+      source_id: sourceReuters.id,
+      highlight_id: hl3.id,
+      context_snippet:
+        "Northgate shares fell 18% on FDA warning letter disclosure.",
+    },
+  });
+
+  await client.mention.create({
+    data: {
+      entity_id: entityHarrington.id,
+      source_id: sourceShareholderLetter.id,
+      context_snippet:
+        "CEO David Harrington asserts 'industry-leading quality culture' in March 2023 shareholder letter.",
+    },
+  });
+
+  await client.mention.create({
+    data: {
+      entity_id: entityHarrington.id,
+      source_id: sourceMemo.id,
+      highlight_id: hl5.id,
+      context_snippet:
+        "CEO David Harrington named as recipient of June 2022 QA escalation memo.",
+    },
+  });
+
+  // ─── Events ───────────────────────────────────────────────────────────────────
+
+  await client.event.create({
+    data: {
+      dossier_id: dossier.id,
+      title: "FDA CGMP Inspection — Raleigh Facility",
+      description:
+        "FDA investigators conducted an unannounced CGMP inspection of Northgate's Raleigh sterile fill facility, spanning 13 business days.",
+      event_date: new Date("2023-10-16"),
+      precision: "day",
+      confidence: 95,
+      claim_id: claimSystemic.id,
+      highlights: { create: [{ highlight_id: hl2.id }] },
+      entities: {
+        create: [{ entity_id: entityNorthgate.id }],
+      },
+      tags: {
+        create: [{ tag_id: tagRegulatory.id }],
+      },
+    },
+  });
+
+  await client.event.create({
+    data: {
+      dossier_id: dossier.id,
+      title: "FDA Issues Warning Letter WL-2024-0108",
+      description:
+        "FDA formally issued warning letter citing three CGMP violations, including 47 instances of audit log backdating.",
+      event_date: new Date("2024-01-08"),
+      precision: "day",
+      confidence: 98,
+      claim_id: claimSystemic.id,
+      highlights: { create: [{ highlight_id: hl1.id }] },
+      entities: {
+        create: [{ entity_id: entityNorthgate.id }],
+      },
+      tags: {
+        create: [
+          { tag_id: tagRegulatory.id },
+          { tag_id: tagManufacturing.id },
+        ],
+      },
+    },
+  });
+
+  await client.event.create({
+    data: {
+      dossier_id: dossier.id,
+      title: "Internal QA Escalation — Chu Memo to Wei and Harrington",
+      description:
+        "VP of QA Raymond Chu formally escalated data integrity concerns in writing to CSO Dr. Sandra Wei and CEO David Harrington, requesting a CAPA within 30 days.",
+      event_date: new Date("2022-06-14"),
+      precision: "day",
+      confidence: 88,
+      claim_id: claimKnowledge.id,
+      highlights: {
+        create: [{ highlight_id: hl4.id }, { highlight_id: hl5.id }],
+      },
+      entities: {
+        create: [{ entity_id: entityWei.id }, { entity_id: entityChu.id }],
+      },
+      tags: {
+        create: [{ tag_id: tagManufacturing.id }],
+      },
+    },
+  });
+
+  await client.event.create({
+    data: {
+      dossier_id: dossier.id,
+      title: "CEO Asserts 'Industry-Leading Quality Culture' in Annual Letter",
+      description:
+        "David Harrington's 2023 annual shareholder letter characterizes the Raleigh facility as operating under an industry-leading quality culture — a claim issued 12 months after Chu's internal escalation.",
+      event_date: new Date("2023-03-22"),
+      precision: "day",
+      confidence: 75,
+      claim_id: claimKnowledge.id,
+      entities: {
+        create: [
+          { entity_id: entityHarrington.id },
+          { entity_id: entityNorthgate.id },
+        ],
+      },
+      tags: {
+        create: [{ tag_id: tagFinancial.id }],
+      },
+    },
+  });
+
+  // ─── Brief ────────────────────────────────────────────────────────────────────
+  // Inline citation tokens ([[cite:highlight:<id>]] / [[cite:source:<id>]])
+  // are rendered as chips by the brief editor and resolve into footnotes at
+  // export time. Seeding them here means the demo brief is evidence-linked on
+  // first open rather than a plain-markdown placeholder.
+
+  const cH = (id: string) => `[[cite:highlight:${id}]]`;
+  const cS = (id: string) => `[[cite:source:${id}]]`;
+
+  await client.brief.create({
+    data: {
+      dossier_id: dossier.id,
+      title: "Northgate Pharma FDA Warning Letter — Research Summary",
+      status: "draft",
+      version: 1,
+      body_markdown: `# Northgate Pharma — FDA Warning Letter Investigation
+
+**Status:** Draft
+**Prepared by:** Alex Mercer
+**Last updated:** March 2024
+
+---
+
+## Summary of Findings
+
+Northgate Pharmaceutical Inc. received FDA Warning Letter WL-2024-0108 on January 8, 2024, citing three CGMP violations at its Raleigh, NC sterile fill facility ${cS(sourceFdaLetter.id)}. The most significant finding — 47 instances of audit log backdating over 18 months ${cH(hl1.id)} — is not an isolated compliance lapse but evidence of a systemic data integrity failure.
+
+Evidence suggests executive leadership was aware of data integrity risks at least 21 months before the FDA inspection, based on an internal QA escalation memo addressed directly to the CSO and CEO in June 2022 ${cH(hl5.id)}.
+
+---
+
+## Key Claims
+
+### 1. Violations Were Systemic (High Confidence — 85%)
+
+The 47 backdating instances, sustained from March 2022 to September 2023, indicate organizational tolerance rather than isolated misconduct ${cH(hl6.id)}. Former FDA investigator Dr. Marcus Obi characterizes the pattern as a management failure. The company's prior Form 483 at the same facility in 2020 ${cH(hl3.id)} suggests the corrective action was inadequate.
+
+### 2. Executive Knowledge Pre-Dated Inspection by 21+ Months (High Confidence — 90%)
+
+A June 2022 internal QA memo from VP Raymond Chu to CSO Dr. Sandra Wei and CEO David Harrington explicitly names data integrity concerns on Sterile Fill Line 3 and requests a CAPA ${cH(hl5.id)}. An earlier escalation to the site Quality Director was made in March 2022 with no corrective action issued ${cH(hl4.id)}.
+
+### 3. Financial Exposure Likely Exceeds $40M Floor (Open — 65%)
+
+Analyst estimates range from $40M to $120M ${cS(sourceReuters.id)}. Dr. Obi advises against anchoring to the lower estimate given the data integrity findings, the prior 483, and the likelihood of a consent decree ${cH(hl7.id)}. **Further research needed:** comparable consent decree financials.
+
+---
+
+## Entities of Interest
+
+| Entity | Role |
+|---|---|
+| Northgate Pharmaceutical Inc. | Subject of investigation |
+| David Harrington | CEO; named recipient of June 2022 escalation memo; author of March 2023 shareholder letter |
+| Dr. Sandra Wei | CSO; named recipient of June 2022 escalation memo |
+| Raymond Chu | VP QA; author of escalation memo; key witness |
+
+---
+
+## Open Questions
+
+- Was the 30-day CAPA requested in the June 2022 memo ever issued?
+- What were the findings of the 2020 Form 483 and what corrective actions were documented?
+- Has Raymond Chu retained separate legal counsel?
+- How does the CEO's March 2023 "industry-leading quality culture" assertion interact with the contemporaneous internal escalation (potential securities disclosure issue)?
+
+---
+
+*This brief is a working document. Claims marked Open require additional source verification.*`,
+    },
+  });
+
+  return { dossierId: dossier.id };
+}
